feat: 애플 결제 전략 정의

+++ src/main/java/com/bookbla/americano/domain/payment/enums/PaymentTable.java
@@ -1,5 +1,6 @@
 package com.bookbla.americano.domain.payment.enums;
 
+import java.math.BigDecimal;
 import java.util.Arrays;
 
 import com.bookbla.americano.base.exception.BaseException;
@@ -11,14 +12,14 @@ import lombok.RequiredArgsConstructor;
 @Getter
 public enum PaymentTable {
 
-    BOOKMARKS_10("bookmarks_10", 2_500, 10),
-    BOOKMARKS_35("bookmarks_35", 7_000, 35),
-    BOOKMARKS_80("bookmarks_80", 14_400, 80),
-    BOOKMARKS_150("bookmarks_150", 22_500, 150),
+    BOOKMARKS_10("bookmarks_10", BigDecimal.valueOf(2_500), 10),
+    BOOKMARKS_35("bookmarks_35", BigDecimal.valueOf(7_000), 35),
+    BOOKMARKS_80("bookmarks_80", BigDecimal.valueOf(14_400), 80),
+    BOOKMARKS_150("bookmarks_150", BigDecimal.valueOf(22_500), 150),
     ;
 
     private final String productId;
-    private final int price;
+    private final BigDecimal price;
     private final int count;
 
     public static PaymentTable from(String productId) {
